+++ bin.ts
@@ -2,6 +2,7 @@
 import { DEFAULT_URL, Integration, Platform } from './lib/Constants';
 import { run } from './lib/Setup';
 import { runNextjsWizard } from './src/nextjs/nextjs-wizard';
+import { runSourcemapsWizard } from './src/sourcemaps/sourcemaps-wizard';
 import { runSvelteKitWizard } from './src/sveltekit/sveltekit-wizard';
 export * from './lib/Setup';
 
@@ -62,6 +63,9 @@ if (argv.i === 'nextjs') {
 } else if (argv.i === 'sveltekit') {
   // eslint-disable-next-line no-console
   runSvelteKitWizard({ promoCode: argv['promo-code'] }).catch(console.error);
+} else if (argv.i === 'sourcemaps') {
+  // eslint-disable-next-line no-console
+  runSourcemapsWizard();
 } else {
   // eslint-disable-next-line @typescript-eslint/no-floating-promises
   run(argv);