+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Add single tenant to self-hosted question (#277)
+
 ## 3.2.3
 
 fix(sveltekit): Bump magicast to handle satisfies keyword (#279)