+++ CHANGELOG.md
@@ -2,12 +2,13 @@
 
 ## Unreleased
 
-- feat: Add empty sourcemaps wizard (#295)
-- feat: Add single tenant to self-hosted question (#277)
-- feat: Improve error handling of incorrect self-hosted URLs (#299)
 - feat(sourcemaps): Add Login and Project Selection flow (#300)
 - feat(sourcemaps): Add Sourcemaps as selectable integration (#302)
 - feat(sourcemaps): Add bundler selection prompt (#304)
+- feat: Add empty sourcemaps wizard (#295)
+- feat: Add single tenant to self-hosted question (#277)
+- feat: Improve error handling of incorrect self-hosted URLs (#299)
+- fix: Add select with sliding window for project selection prompt #306
 
 ## 3.2.3