+++ CHANGELOG.md
@@ -2,9 +2,9 @@
 
 ## Unreleased
 
+- feat(sourcemaps): Add bundler selection prompt (#304)
 - feat(sourcemaps): Add Login and Project Selection flow (#300)
 - feat(sourcemaps): Add Sourcemaps as selectable integration (#302)
-- feat(sourcemaps): Add bundler selection prompt (#304)
 - feat: Add empty sourcemaps wizard (#295)
 - feat: Add single tenant to self-hosted question (#277)
 - feat: Improve error handling of incorrect self-hosted URLs (#299)