+++ CHANGELOG.md
@@ -8,7 +8,8 @@
 - feat: Add empty sourcemaps wizard (#295)
 - feat: Add single tenant to self-hosted question (#277)
 - feat: Improve error handling of incorrect self-hosted URLs (#299)
-- fix: Add select with sliding window for project selection prompt #306
+- fix: Add select with sliding window for project selection prompt (#306)
+- feat: Add telemetry helper (#309)
 
 ## 3.2.3